feat: Refactor typia usage and improve OpenAI response
This commit refactors the usage of typia and improves the OpenAI
response. The import of typia is cleaned up and a new utility function
`typiaJsonToOpenAIResponse` is introduced. This function converts a
typia JSON schema to an OpenAI response format. The OpenAI chat
completion's response format is also updated to use this new function.